Generated 5000 datasets of t.test for values generated by normal distribution with sample size 30 and standard deviation 5.
This is repeated for mean 0 to 6.

The p.value and the estimate of the mean is recorded to the dataframe.

Created a plot of the relation between the true mean vs proportion of the values that the null hypothesis is rejected with alpha = 0.05.

Performed small analysis of the relationship between effect size and the power of the test.